requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
click==8.3.1
fastapi==0.135.2
greenlet==3.3.2
h11==0.16.0
idna==3.11
joblib==1.5.3
numpy==2.4.4
pandas==3.0.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
RapidFuzz==3.14.3
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==1.0.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
